[refactor] 패널티 api 출석 api에 병합
